requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
google-generativeai>=0.8.0
flask>=2.3.0
flask-cors>=4.0.0
nltk>=3.8.0
numpy>=1.24.0
safetensors>=0.3.0
wikipedia-api>=0.6.0
python-dotenv>=1.0.0
